pyproject.toml
[tool.poetry]
name = "python-structure"
version = "0.1.0"
description = "My python projects structure"
authors = ["rmenai <rami.menai@outlook.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"

[tool.poetry.dev-dependencies]
python-dotenv = "^0.19.0"
pytest = "^6.2.5"
colorlog = "^6.5.0"
taskipy = "^1.8.2"
pre-commit = "^2.15.0"
flake8 = "^3.9.2"
flake8-annotations = "^2.6.2"
flake8-bugbear = "^21.9.1"
flake8-docstrings = "^1.6.0"
flake8-import-order = "^0.18.1"
flake8-string-format = "^0.3.0"
flake8-tidy-imports = "^4.4.1"
flake8-todo = "^0.7"
pep8-naming = "^0.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
start = "python -m app"
tests = "pytest app/tests/"
lint = "pre-commit run --all-files"
precommit = "pre-commit install"
